feat(programs): add early zen-browser flake configuration
Add early Zen Browser flake configuration to replace the initial
Flathub version soon.
Still need to figure out how to make profiles work. zenix could be a
simple solution if it works but could delay Zen updates significantly
as they would depend on the updates of zenix and the flake to package
Zen that it depends on, if I am understanding it correctly.
Profiles should also allow to set default settings for extensions
according to this list of options
https://home-manager-options.extranix.com/?query=settings+firefox&release=master