feat(gamemode): remove volume adjustment
Remove automatic volume adjustment for gamemode as it is highly
dependent on the game played and thus it does not make sense to have a
unique setting for it.
Also, the volume can be easily controlled with keyboard commands
instead without having to alt-tab, which helps circumvent the initial
issue of not having a working Stream Deck setup to modulate it.