fix(input): restore dead keys functionality on gtk apps
Since version 4.20, GTK is no longer involved in handling compose and
dead keys on Wayland, as it should be the responsibility of the input
method and not of the graphical toolkit.

A more comprehensible explanation of the issue is available here:
https://github.com/ghostty-org/ghostty/discussions/8899#discussioncomment-14717979

The two most popular input method frameworks right now are IBus and
Fcitx5.
IBus works to restore the dead keys functionality but sets the
GTK_IM_MODULE and QT_IM_MODULE environment variables which triggers a
warning when logging in as it might cause some issues. For some
reason, there does not seem to be an option in the module to disable
this behaviour.
This is not an issue with Fcitx5 which provides the waylandFrontend
option to disable setting these variables.